fix(DeltaLog): ensure posts with multiple links don't extract invalid link contents #132
